src/main/webapp/app/entities/contrat-abonnement/delete/contrat-abonnement-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, fakeAsync, inject, tick } from '@angular/core/testing';
import { HttpResponse, provideHttpClient } from '@angular/common/http';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { ContratAbonnementService } from '../service/contrat-abonnement.service';

import { ContratAbonnementDeleteDialogComponent } from './contrat-abonnement-delete-dialog.component';

describe('ContratAbonnement Management Delete Component', () => {
  let comp: ContratAbonnementDeleteDialogComponent;
  let fixture: ComponentFixture<ContratAbonnementDeleteDialogComponent>;
  let service: ContratAbonnementService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [ContratAbonnementDeleteDialogComponent],
      providers: [provideHttpClient(), NgbActiveModal],
    })
      .overrideTemplate(ContratAbonnementDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(ContratAbonnementDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(ContratAbonnementService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
